notedeck_ui: move carousel to `note/media.rs`
